FIX: correct the type

+++ forecaster/data/utils.py
@@ -57,7 +57,7 @@ def create_bin_labels(df, column, bins):
     """Create column labels based on bins for a DataFrame starting with 0"""
     labels = pd.cut(df[column], bins=bins, labels=False)
     max_label = labels.max()
-    return labels.fillna(max_label + 1)
+    return labels.fillna(max_label + 1).astype(int)
 
 
 def create_spatial_labels_kmeans(dataframe, num_clusters):
